CSN: CSN-platforms-494
* Applying Andrey Eefimov's patch to fix UMF_CHUNK_BYTES when set to 8.

+++ modules/bluespec/common/fpgaenv/physical-platform/physical-devices/unix-pipe/unix-pipe-device-bdpi.c
@@ -380,7 +380,7 @@ void pipe_write(unsigned char handle, unsigned long long data)
 {   
     int bytes_written;
     unsigned char databuf[BDPI_CHUNK_BYTES];
-    unsigned int mask;
+    unsigned long long mask;
     int i;
     Channel *channel;
